[Fix #10971] Add `diff_comma` option to Style/TrailingCommaInArguments
Introduce a new EnforcedStyleForMultiline, `diff_comma`, for
Style/TrailingCommaInArguments. `diff_comma` requires a trailing comma
only when the final argument is followed by an immediate newline. When
the closing `)` or `]` shares a line with the last argument, a trailing
comma is not allowed.

Examples:

  ```ruby
  # good under diff_comma, bad under consistent_comma
  some_method(a: "b",
              c: "d")

  foo(1, [
    2,
  ])

  # bad under diff_comma, good under consistent_comma
  some_method(a: "b",
              c: "d",)

  foo(1, [
    2,
  ],)
  ```

In line with #13806, this adds a separate style rather than modifying
`consistent_comma` to preserve backward compatibility.

Most of the required functionality was already implemented in #13806,
so this PR largely reuses that implementation. The only gap was support
for the `[]` method; this PR adapts it with minor changes to handle
`[]`.