[Fix #14486] Fix false positives for `Style/RedundantParentheses` with unary operators and `yield`, `super`, or `defined?`
Removing parens for `defined?` does not result in a syntax error
but because of it's low precedence it's better to keep them anyways.
For example, in `(defined? foo) && baz`, removing the parens makes it into a single `defined?` invocation.
Since code like `(!yield arg)` should continue to register an offense, this also updates normal method calls
to do the same. As such, I added a few more testcases for that (+ a changelog entry)