Gracefully handle invalid configs during LSP addon startup
If the config is invalid for any reason during startup, the addon will be left in a state
which it cannot recover from.
This fixes this by instead showing a notification to the user.
After the config is fixed, it will be properly reloaded.
While it is invalid, no diagnostics will be shown.

I have also:
* Removed reliance on `RubyLsp::Requests::Support::InternalRuboCopError`
* Moved lsp specific error handling into the addon itself
* Logged the full error

The main goal was to gracefully handle editing the config,
where it may be invalid shortly, however due to a ruby-lsp bug no feedback is given in that case yet.