Fix the built-in Ruby LSP add-on not respecting `.rubocop` config file
If the user specifies `-c foo.yml`, it should be used.

This consolidates option/cli handling to `apply_options!`.
That way `force_default_config` is consistenly used as well.

That now also happens when generating a todo. It seems correct for consistency, but I didn't add a test for it and am ok with dropping that change.