[Fix #14516] Add `AllowImplicitArrayBrackets` to `Layout/FirstArrayElementLineBreak`
This PR adds `AllowImplicitArrayBrackets` to `Layout/FirstArrayElementLineBreak`.
To prioritize compatibility with existing implicit array brackets behavior,
the default is `AllowImplicitArrayBrackets: false`.

Note that an assignment such as `var = key: value`, where hash braces are omitted,
is invalid syntax and results in an error. Therefore, adding an equivalent option to
`Layout/FirstHashyElementLineBreak` is unnecessary.

Fixes #14516.