Add FloatPrecision cop
The FloatPrecision cop catches float literals that cannot be precisely
represented by an IEEE 754 double presision floating point number. This
behaviour can be surprising, so we have other cops (e.g. FloatComparison)
that prevents surprises like this.

This cop catches float literals that cannot be precisely represented by
a literal (e.g. 100000000000.000000000001).

The implementation of this cop is a bit complex, but can be greatly
simplfied if we are allowed to depend on BigDecimal. However, BigDecimal
is no longer a default gem since Ruby 3.3 so we will need to add it as
a dependency.