[Fix #14552] Improve `Security/JSONLoad` documentation and implementation
Things have somewhat changed since this cop was added. In addition, it should not register an offense
when `create_additions` is explicitly passed.
The user has made a concious decision in that case.

Removed docs for `quirks_mode`, that has been gone since 2.0.0 from the 2015. It just works now, no need
to bother users with this anymore.