Handle non-specific issues with the Gemfile to allow fallback
In general, when the Gemfile fails to load for any reason, we want to
fallback to try loading from the system gems.

A non-specific issue could be something like eval_gemfile failing when
a submodule that includes the requisite file has not yet been updated.

This commit changes the rescue to the generic StandardError, which
subsumes the Bundler errors, and handles other errors as well.