[Fix #14628] Fix false positives for `Style/FloatDivision`
This PR fixes false positives for `Style/FloatDivision`
when using `Regexp.last_match` or nth reference (e.g., `$1`).

For `Regexp.last_match(1)` and `$1`, it assumes that the value is a string
matched by a regular expression, and allows conversion with `#to_f`.

This cop is already marked as unsafe, but these issues can be avoided.

Fixes #14628.