[Fix #10147] Make `Lint/ElseLayout` allow single-line `else` body in `then` conditional
This PR makes `Lint/ElseLayout` allow a single-line `else` body in `then` single-line conditional.

For single-line conditionals using `then` the layout is disallowed
when the `else` body is multiline because it is treated as a lint offense.

```ruby
# bad
if something then on_the_same_line_as_then
else first_line
  second_line
end
```

For single-line conditionals using `then` the layout is allowed
when the `else` body is a single-line because it is treated as intentional.

```ruby
# good
if something then on_the_same_line_as_then
else single_line
end
```

Resolves #10147.