Allow disabling cache cleanup
In our use case, we have CI nodes that are recycled frequently,
none live more than a say, so cleaning up the cache directory
is just a waste of time.

I could set the limit to a very large number, but that is worse
because then rubocop will keep getting slower counting cache entries
just to realize there's nothing to cleanup.